fix supporter routing

+++ backend/IntexAPI/IntexAPI/Data/Donation.cs
@@ -36,10 +36,10 @@ public class Donation
     public string? CurrencyCode { get; set; }
 
     [Column("amount")]
-    public decimal? Amount { get; set; }
+    public double? Amount { get; set; }
 
     [Column("estimated_value")]
-    public decimal? EstimatedValue { get; set; }
+    public double? EstimatedValue { get; set; }
 
     [Column("impact_unit")]
     [MaxLength(64)]
@@ -49,5 +49,5 @@ public class Donation
     public string? Notes { get; set; }
 
     [Column("referral_post_id")]
-    public int? ReferralPostId { get; set; }
+    public double? ReferralPostId { get; set; }
 }
